Moved GameOption class from 'enums' to 'account_info' module
