+++ hearthy/protocol/enums.py
@@ -441,121 +441,3 @@ class TrackWhat(IntEnum):
     TRACK_CM_SHOW_PREMIUMS_NOT_OWNED = 37
     TRACK_CM_HIDE_PREMIUMS_NOT_OWNED = 38
     TRACK_PLAY_FORGE_QUEUE = 39
-
-class GameOption(IntEnum):
-    AI_MODE = 0x30
-    BACKGROUND_SOUND = 20
-    BUNDLE_JUST_PURCHASE_IN_HUB = 0x6c
-    CARD_BACK = 0x1c
-    CARD_BACK2 = 0x1d
-    CHANGED_CARDS_DATA = 0x24
-    CONNECT_TO_AURORA = 0x1f
-    COVER_MOUSE_OVERS = 0x2f
-    CURSOR = 3
-    DECK_PICKER_MODE = 0x36
-    FAKE_PACK_COUNT = 14
-    FAKE_PACK_OPENING = 13
-    FRIENDS_LIST_CURRENTGAME_SECTION_HIDE = 0x3a
-    FRIENDS_LIST_FRIEND_SECTION_HIDE = 0x3b
-    FRIENDS_LIST_NEARBYPLAYER_SECTION_HIDE = 60
-    FRIENDS_LIST_RECRUIT_SECTION_HIDE = 0x3d
-    FRIENDS_LIST_REQUEST_SECTION_HIDE = 0x39
-    GFX_FULLSCREEN = 10
-    GFX_FXAA = 0x19
-    GFX_HEIGHT = 9
-    GFX_MSAA = 0x18
-    GFX_QUALITY = 12
-    GFX_TARGET_FRAME_RATE = 0x1a
-    GFX_VSYNC = 0x1b
-    GFX_WIDTH = 8
-    GFX_WIN_CAMERA_CLEAR = 0x17
-    GFX_WIN_POSX = 0x26
-    GFX_WIN_POSY = 0x27
-    HAS_ACKED_ARENA_REWARDS = 0x66
-    HAS_ADDED_CARDS_TO_DECK = 0x58
-    HAS_BEEN_NUDGED_TO_CM = 0x57
-    HAS_CLICKED_TOURNAMENT = 0x3e
-    HAS_CRAFTED = 0x5e
-    HAS_DISENCHANTED = 0x5b
-    HAS_ENTERED_NAXX = 0x6a
-    HAS_FINISHED_A_DECK = 70
-    HAS_LOST_IN_ARENA = 100
-    HAS_OPENED_BOOSTER = 0x3f
-    HAS_PLAYED_EXPERT_AI = 90
-    HAS_PLAYED_NAXX = 0x6d
-    HAS_RUN_OUT_OF_QUESTS = 0x65
-    HAS_SEEN_100g_REMINDER = 0x61
-    HAS_SEEN_ALL_BASIC_CLASS_CARDS_COMPLETE = 0x56
-    HAS_SEEN_BRM = 0x73
-    HAS_SEEN_CINEMATIC = 11
-    HAS_SEEN_COLLECTIONMANAGER = 0x41
-    HAS_SEEN_COLLECTIONMANAGER_AFTER_PRACTICE = 0x72
-    HAS_SEEN_CRAFTING_INSTRUCTION = 0x5d
-    HAS_SEEN_CUSTOM_DECK_PICKER = 0x55
-    HAS_SEEN_DECK_HELPER = 0x52
-    HAS_SEEN_EXPERT_AI = 80
-    HAS_SEEN_EXPERT_AI_UNLOCK = 0x51
-    HAS_SEEN_FORGE = 0x47
-    HAS_SEEN_FORGE_1WIN = 0x4c
-    HAS_SEEN_FORGE_2LOSS = 0x4d
-    HAS_SEEN_FORGE_CARD_CHOICE = 0x49
-    HAS_SEEN_FORGE_CARD_CHOICE2 = 0x4a
-    HAS_SEEN_FORGE_HERO_CHOICE = 0x48
-    HAS_SEEN_FORGE_PLAY_MODE = 0x4b
-    HAS_SEEN_FORGE_RETIRE = 0x4e
-    HAS_SEEN_GOLD_QTY_INSTRUCTION = 0x62
-    HAS_SEEN_HEROIC_WARNING = 0x68
-    HAS_SEEN_HUB = 0x45
-    HAS_SEEN_LEVEL_3 = 0x63
-    HAS_SEEN_MULLIGAN = 0x4f
-    HAS_SEEN_NAXX = 0x69
-    HAS_SEEN_NAXX_CLASS_CHALLENGE = 0x6b
-    HAS_SEEN_PACK_OPENING = 0x53
-    HAS_SEEN_PRACTICE_MODE = 0x54
-    HAS_SEEN_PRACTICE_TRAY = 0x44
-    HAS_SEEN_SHOW_ALL_CARDS_REMINDER = 0x5c
-    HAS_SEEN_STEALTH_TAUNTER = 0x67
-    HAS_SEEN_THE_COIN = 0x60
-    HAS_SEEN_TOURNAMENT = 0x40
-    HAS_STARTED_A_DECK = 0x71
-    HEALTHY_GAMING_DEBUG = 15
-    HUD = 4
-    IDLE_KICK_TIME = 0x13
-    IDLE_KICKER = 0x12
-    IN_RANKED_PLAY_MODE = 0x5f
-    INTRO = 0x2c
-    INVALID = 0
-    JUST_FINISHED_TUTORIAL = 0x42
-    KELTHUZADTAUNTS = 0x25
-    LAST_CUSTOM_DECK_CHOSEN = 0x35
-    LAST_FAILED_DOP_VERSION = 0x29
-    LAST_PRECON_HERO_CHOSEN = 0x34
-    LAST_SCENE_MODE = 0x10
-    LAST_SELECTED_STORE_ADVENTURE_ID = 0x70
-    LAST_SELECTED_STORE_BOOSTER_ID = 0x6f
-    LOCAL_TUTORIAL_PROGRESS = 30
-    LOCALE = 0x11
-    MUSIC = 2
-    MUSIC_VOLUME = 7
-    NEARBY_PLAYERS = 0x16
-    PAGE_MOUSE_OVERS = 0x2e
-    PREFERRED_CDN_INDEX = 40
-    PREFERRED_REGION = 0x15
-    RECONNECT = 0x21
-    RECONNECT_RETRY_TIME = 0x23
-    RECONNECT_TIMEOUT = 0x22
-    SEASON_END_THRESHOLD = 0x20
-    SELECTED_ADVENTURE = 0x37
-    SELECTED_ADVENTURE_MODE = 0x38
-    SHOW_ADVANCED_COLLECTIONMANAGER = 0x43
-    SHOWN_GFX_DEVICE_WARNING = 0x2b
-    SOUND = 1
-    SOUND_VOLUME = 6
-    SPECTATOR_OPEN_JOIN = 110
-    STREAMING = 5
-    TIP_CRAFTING_UNLOCKED = 0x59
-    TIP_FORGE_PROGRESS = 0x33
-    TIP_PLAY_PROGRESS = 50
-    TIP_PRACTICE_PROGRESS = 0x31
-    TOUCH_MODE = 0x2a
-    TUTORIAL_LOST_PROGRESS = 0x2d
